put '/' into variable for code to be easier to read. put context provider to the seperate component outside app.tsx
