+++ src/pages/EditView/index.tsx
@@ -12,7 +12,7 @@ import { Typography } from "@mui/material";
 import { useParams } from "react-router-dom";
 import { useContext } from "react";
 import { RecipeContextType } from "../../shared/types/RecipeContextType.ts";
-import { RecipeContext } from "../../App.tsx";
+import { RecipeContext } from "../../shared/components/RecipeContextProvider";
 import { Recipe } from "../../shared/types/Recipe.ts";
 
 type ParamsInterface = {
